requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
pytest==7.4.3 